@@ -1,36 +1,49 @@
 {
-  "name": "react-node-graph",
-  "version": "1.1.1",
+  "name": "naph",
+  "version": "0.0.0",
   "description": "UI for rendering Node Graphs.",
-  "main": "lib/index.js",
+  "main": "build/index.js",
   "peerDependencies": {
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0"
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0"
   },
   "devDependencies": {
+    "@auto-it/conventional-commits": "^7.13.1",
+    "auto": "^7.13.1",
+    "tslib": "^1.10.0",
+    "@zeit/ncc": "^0.20.5",
+    "@types/lodash.clonedeep": "^4.5.6",
+    "@types/react": "^16.9.16",
+    "@types/react-onclickoutside": "^6.7.3",
+    "@typescript-eslint/eslint-plugin": "^2.11.0",
+    "@typescript-eslint/parser": "^2.11.0",
     "babel": "^5.6.14",
     "babel-browser-transform": "^0.1.0",
     "babel-core": "^5.6.18",
     "babel-eslint": "^4.1.3",
     "babel-loader": "^5.3.1",
-    "react": "^16.9.0",
-    "react-hot-loader": "^1.2.8",
+    "eslint-plugin-react": "^7.12.4",
+    "eslint-plugin-react-hooks": "^1.6.0",
+    "react": "^16.12.0",
+    "react-hot-loader": "^4.12.18",
+    "typescript": "^3.7.3",
     "webpack": "^1.10.1",
     "webpack-dev-server": "^1.10.1"
   },
   "dependencies": {
     "lodash.clonedeep": "^4.5.0",
-    "react-dom": "^16.9.0",
+    "react-dom": "^16.12.0",
     "react-draggable": "^3.3.2",
     "react-onclickoutside": "^6.9.0"
   },
   "scripts": {
-    "build": "babel src --out-dir lib && webpack --config webpack.prod.config.js",
-    "dev": " webpack-dev-server --progress --display-errors-details --inline --hot --content-base --port 9999"
+    "tsc": "tsc -d",
+    "build": "yarn tsc",
+    "release": "auto shipit"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/lightsinthesky/react-node-graph.git"
+    "url": "https://github.com/williamluke4/react-node-graph.git"
   },
   "keywords": [
     "react",
@@ -40,6 +53,6 @@
     "graph",
     "ui"
   ],
-  "author": "Andrew Wilson",
+  "author": "William Luke",
   "license": "MIT"
 }
